fix: Consolidate duplicate patch-logger
