+++ packages/functions/src/auth/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./discord"
 export * from "./github"
-export * from "./patch-logger"
