+++ packages/functions/src/utils/patch-logger.ts
